Re-audit against BuddyDoom's legacy fixes; stop hiding C4013
Re-ran the §9 audit now that Windows builds exist, because Windows x64 is
LLP64 (long is 4 bytes, pointers 8) while the Linux build is LP64 -- code
can be LP64-clean and still break on Windows, which is exactly how
BuddyDoom's worst 64-bit bug surfaced.

Nothing was missing. Every item on their list is either already present
through MBF/killough heritage or already done here: MAXSEGS is
MAX_SCREENWIDTH/2+1, drawsegs grows by doubling, the zone is
malloc-backed, texturecolumnofs is already 32-bit with no >64k composite
guard, MAXSWITCHES is dynamic, and snd_channels defaults to 32 rather
than the vanilla 3.

Their §7 (realloc moving an array the zone holds owner back-pointers
into) is structurally impossible here: SMMU keeps each lump's cache
pointer inside its lumpinfo_t rather than in a parallel lumpcache[], so
growing the lump list at runtime cannot dangle a back-pointer.

Their §1 crash -- an implicitly-declared calloc returning int, truncating
a 64-bit pointer -- was checked mechanically rather than by eye: built all
91 files with MSVC's C4013 enabled. Zero implicit declarations.

Two things that check did turn up, both fixed:

- Makefile.msvc suppressed C4013. That is the one warning that catches the
  above, and suppressing it would have hidden the same bug had it ever
  been introduced. Removed, with a comment on why it stays off. It costs
  nothing today since the tree is clean.
- m_misc.c printed intptr_t with %d in the config-help output. Benign on
  x64 varargs but wrong; narrowed explicitly.

+++ m_misc.c
@@ -1813,12 +1813,15 @@ void M_SaveDefaults (void)
 	     fprintf(f,"[(\"%s\")]", (char *) dp->defaultvalue) :
 	     dp->limit.min == UL ?
 	     dp->limit.max == UL ?
-	     fprintf(f, "[?-?(%d)]", dp->defaultvalue) :
-	     fprintf(f, "[?-%d(%d)]", dp->limit.max, dp->defaultvalue) :
+	     // defaultvalue is intptr_t (it doubles as a char * for string
+	     // defaults), so it must be narrowed explicitly for %d rather than
+	     // passed 64 bits wide into a 32-bit conversion
+	     fprintf(f, "[?-?(%d)]", (int) dp->defaultvalue) :
+	     fprintf(f, "[?-%d(%d)]", dp->limit.max, (int) dp->defaultvalue) :
 	     dp->limit.max == UL ?
-	     fprintf(f, "[%d-?(%d)]", dp->limit.min, dp->defaultvalue) :
+	     fprintf(f, "[%d-?(%d)]", dp->limit.min, (int) dp->defaultvalue) :
 	     fprintf(f, "[%d-%d(%d)]", dp->limit.min, dp->limit.max,
-		     dp->defaultvalue)) == EOF ||
+		     (int) dp->defaultvalue)) == EOF ||
 	    fprintf(f," %s %s\n", dp->help, dp->wad_allowed ? "*" :"") == EOF)
 	  goto error;
